backend/requirements.txt
Flask
firebase-admin
google-genai
pandas
clarifai
joblib
scikit-learn
flask_cors
python-dotenv